Move API.ai objects and updaters into their own namespaces
OS-11
